+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - ELR Parser
     - Export `ELR.commit` and `ELR.rollback`.
     - **Breaking Change**: Reducer will take `ReducerContext` as the parameter instead of `ParserContext`.
+    - **Breaking Change**: `ParserBuilder.checkSymbols/checkAll` will require a lexer to check if literals in grammar rules are valid.
 - Optimize performance.
 
 ## v0.4.0